+++ frontend/src/features/game/routes/Matching.tsx
@@ -1,0 +1,40 @@
+import { memo, FC } from 'react';
+import {
+  Box,
+  Center,
+  Divider,
+  Flex,
+  Heading,
+  Spinner,
+  Stack,
+} from '@chakra-ui/react';
+import { PrimaryButton } from 'components/atoms/button/PrimaryButton';
+import { ContentLayout } from 'components/ecosystems/ContentLayout';
+import { MatchState, useGameMatching } from '../hooks/useGameMatching';
+
+export const Matching: FC = memo(() => {
+  const { setMatchState } = useGameMatching();
+
+  const onClickCancel = () => {
+    setMatchState(MatchState.MatchingCancel);
+  };
+
+  return (
+    <ContentLayout title="">
+      <Center>
+        <Flex align="center" justify="center" height="40vh">
+          <Box bg="white" w="sm" p={4} borderRadius="md" shadow="md">
+            <Heading as="h1" size="lg" textAlign="center">
+              Now Matching...
+            </Heading>
+            <Divider />
+            <Stack spacing={4} py={4} px={10} align="center">
+              <Spinner />
+              <PrimaryButton onClick={onClickCancel}>Cancel</PrimaryButton>
+            </Stack>
+          </Box>
+        </Flex>
+      </Center>
+    </ContentLayout>
+  );
+});
