+++ frontend/src/features/game/routes/Top.tsx
@@ -1,21 +1,15 @@
 import { memo, FC } from 'react';
-import { Center } from '@chakra-ui/react';
+import { Button, Center } from '@chakra-ui/react';
+import { useNavigate } from 'react-router-dom';
 import { ContentLayout } from 'components/ecosystems/ContentLayout';
-import { GameTop } from '../components/GameTop';
-import { Matching } from '../components/Matching';
-import { MatchState, useGameMatching } from '../hooks/useGameMatching';
 
 export const Top: FC = memo(() => {
-  const { matchState, setMatchState } = useGameMatching();
+  const navigate = useNavigate();
 
   return (
     <ContentLayout title="">
       <Center>
-        {matchState === MatchState.None ? (
-          <GameTop setMatchState={setMatchState} />
-        ) : (
-          <Matching setMatchState={setMatchState} />
-        )}
+        <Button onClick={() => navigate('/app/matching')}>Match</Button>
       </Center>
     </ContentLayout>
   );
