+++ frontend/src/routes/PublicRoutes.tsx
@@ -15,8 +15,9 @@ import { CreateChatRooms } from 'features/chat/routes/CreateChatRooms';
 import { DmRoom } from 'features/dm/routes/DmRoom';
 import { DmRooms } from 'features/dm/routes/DmRooms';
 import { Users } from 'features/friends/routes/Users';
-import { Matching } from 'features/game/components/Matching';
+import { Game } from 'features/game/routes/Game';
 import { Games } from 'features/game/routes/Games';
+import { Matching } from 'features/game/routes/Matching';
 import { Top } from 'features/game/routes/Top';
 import { Profile } from 'features/profile/routes/Profile';
 import { ProfileEdit } from 'features/profile/routes/ProfileEdit';
@@ -79,6 +80,7 @@ export const publicRoutes = [
       { path: 'chat/create', element: <CreateChatRooms /> },
       { path: 'chat/:id', element: <ChatRoom /> },
       { path: 'games', element: <Games /> },
+      { path: 'games/:id', element: <Game /> },
       { path: '', element: <Top /> },
       { path: 'matching', element: <Matching /> },
       { path: 'dm', element: <DmRooms /> },
